aom-qm: Fix overflow in quantizer calculation
By using a combination of a small qmlevel and large qindex,
it is possible for the quantizer 'dqv' to end up > 2^15.
This currently causes dqv to overflow, so widen its type
to avoid that.

Note that this is certainly an extreme case, but not entirely
useless - if using a 32x32 transform, the dequantized value
can be as small as dqv/2, which may still fit in 16 bits.

+++ av1/decoder/detokenize.c
@@ -130,7 +130,7 @@ static int decode_coefs(MACROBLOCKD *xd, PLANE_TYPE type, tran_low_t *dqcoeff,
   const uint8_t *band_translate = get_band_translate(tx_size);
   int dq_shift;
   int v, token;
-  int16_t dqv = dq[0];
+  int32_t dqv = dq[0];
 #if CONFIG_NEW_QUANT
   const tran_low_t *dqv_val = &dq_val[0][0];
 #endif  // CONFIG_NEW_QUANT
